requirements.txt
allure-pytest==2.13.5
pytest==8.2.0
requests==2.31.0
selenium==4.21.0
webdriver-manager==4.0.1
python-dotenv~=1.0.1